requirements.txt
fastapi>=0.115.4
uvicorn>=0.18.2
python-dotenv>=0.19.2
pydantic>=2.9.2
psycopg>=3.2.10
psycopg2>=2.9.3